macos/requirements.txt
sounddevice
soundfile
numpy
openai
pyperclip
pynput
python-dotenv
